% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim_kernel.R
\name{discrim_kernel}
\alias{discrim_kernel}
\title{Kernel discriminant analysis}
\usage{
discrim_kernel(mode = "classification", engine = "ks")
}
\arguments{
\item{mode}{A single character string for the type of model. The only
possible value for this model is "classification".}

\item{engine}{A single character string for the engine.}
}
\description{
`discrim_kernel()` defines a model that estimates a multivariate
 distribution for the predictors separately for the data in each class
 non-parameterically with a kernel density estimator. Bayes' theorem is used
 to compute the probability of each class, given the predictor values. This
 function can fit classification models.

\Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("discrim_kernel")}

More information on how \pkg{parsnip} is used for modeling is at
\url{https://www.tidymodels.org/}.
}
