% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ks_kda_wrapper}
\alias{ks_kda_wrapper}
\title{Wrapper around ks::kda to map y -> x.group and also check that ncol(x) <= 3
For ncol(x) > 3, ks::kda() does not have a separate predict() method, and while
We could hack it in by modifying ks::kda(), it needs a lot of data for more
than 3 dimensions anyway.}
\usage{
ks_kda_wrapper(x, y, ...)
}
\description{
Internal wrapper functions
}
\keyword{internal}
